Better App Engine support.
go: use `goapp` for the `build` and `test` subcommands. These need to be App Engine-aware, and editors are easier to configure if the command has the same name for regular packages and App Engine packages.

errcheck: Prepend `-tags appengine` to command line. Same reason as above, mostly: easier to configure commands if this doesn't need to be explicit.